refactor: componenrt name
