+++ packages/web/src/components/organisms/Withdraw/index.tsx
@@ -3,10 +3,10 @@ import styled from 'styled-components'
 import { useProvider } from 'src/fixtures/wallet/hooks'
 import { getMyStakingAmount } from 'src/fixtures/dev-kit/client'
 import { useGetEthPrice } from 'src/fixtures/uniswap/hooks'
-import { useWithdrawStaking, useGetEstimateGas4WithdrawStakingAmount } from 'src/fixtures/dev-kit/hooks'
+import { useGetEstimateGas4WithdrawStakingAmount, useWithdrawByPosition } from 'src/fixtures/dev-kit/hooks'
 import { toAmountNumber, toNaturalNumber, whenDefinedAll } from 'src/fixtures/utility'
-import { TransactForm } from 'src/components/molecules/TransactForm'
-import { FormContainer } from 'src/components/molecules/TransactForm/FormContainer'
+import { WithdrawTransactForm } from 'src/components/molecules/WithdrawTransactForm'
+import { FormContainer } from 'src/components/molecules/WithdrawTransactForm/FormContainer'
 import { EstimatedGasFeeCard } from 'src/components/molecules/EstimatedGasFeeCard'
 import { InfoCircleOutlined } from '@ant-design/icons'
 import { message } from 'antd'
@@ -36,27 +36,14 @@ const SubtitleContianer = styled.div`
 export const Withdraw = ({ className, title, propertyAddress, onChange: onChangeAmount, isDisplayFee }: Props) => {
   const [withdrawAmount, setWithdrawAmount] = useState<string>('')
   const { web3, accountAddress } = useProvider()
-  const { withdrawStaking } = useWithdrawStaking()
+  const { withdrawByPosition } = useWithdrawByPosition()
   const { estimateGas } = useGetEstimateGas4WithdrawStakingAmount(propertyAddress, withdrawAmount || '0')
   const { data: ethPrice } = useGetEthPrice()
+
+  const amountNumber = useMemo(() => toAmountNumber(withdrawAmount), [withdrawAmount])
   const estimateGasUSD = useMemo(
     () => whenDefinedAll([estimateGas, ethPrice], ([gas, eth]) => gas.multipliedBy(eth)),
     [estimateGas, ethPrice]
-  )
-  const withdrawFor = useCallback(
-    (amount: string) => {
-      if (!web3) {
-        message.warn({ content: 'Please sign in', key: 'WithdrawButton' })
-        return
-      }
-      const amountNumber = toAmountNumber(amount)
-      if (amountNumber.toNumber() < 0) {
-        message.warn({ content: 'Please enter a value greater than or equal to 0', key: 'WithdrawButton' })
-        return
-      }
-      withdrawStaking(propertyAddress, amountNumber)
-    },
-    [withdrawStaking, propertyAddress, web3]
   )
   const onClickMax = () =>
     whenDefinedAll([web3, accountAddress], ([libWeb3, account]) =>
@@ -67,6 +54,20 @@ export const Withdraw = ({ className, title, propertyAddress, onChange: onChange
   const onChange = (event: ChangeEvent<HTMLInputElement>) => {
     setWithdrawAmount(event.target.value)
   }
+  const withdraw = useCallback(
+    (sTokenId: string) => {
+      if (!web3) {
+        message.warn({ content: 'Please sign in', key: 'WithdrawButton' })
+        return
+      }
+      if (amountNumber.toNumber() < 0) {
+        message.warn({ content: 'Please enter a value greater than or equal to 0', key: 'WithdrawButton' })
+        return
+      }
+      withdrawByPosition(sTokenId, amountNumber.toString())
+    },
+    [web3, amountNumber, withdrawByPosition]
+  )
   useEffect(() => {
     if (onChangeAmount) {
       onChangeAmount(withdrawAmount)
@@ -78,13 +79,13 @@ export const Withdraw = ({ className, title, propertyAddress, onChange: onChange
   return (
     <FormContainer>
       <Label />
-      <TransactForm
+      <WithdrawTransactForm
         className={className}
         id="withdraw"
         enterButton="Withdraw"
         value={withdrawAmount}
         onChange={onChange}
-        onSearch={withdrawFor}
+        withdraw={withdraw}
         disabled={!web3}
         onClickMax={onClickMax}
         propertyAddress={propertyAddress}
